+++ src/Symfony/Bundle/FrameworkBundle/CHANGELOG.md
@@ -32,6 +32,7 @@ CHANGELOG
  * Add `--exclude` option to the `cache:pool:clear` command
  * Add parameters deprecations to the output of `debug:container` command
  * Change `framework.asset_mapper.importmap_polyfill` from a URL to the name of an item in the importmap
+ * Provide `$buildDir` when running `CacheWarmer` to build read-only resources
 
 6.3
 ---