+++ src/Symfony/Component/HttpKernel/CHANGELOG.md
@@ -15,6 +15,7 @@ CHANGELOG
  * Support the `!` character at the beginning of a string as a negation operator in the url filter of the profiler
  * Deprecate `UriSigner`, use `UriSigner` from the HttpFoundation component instead
  * Deprecate `FileLinkFormatter`, use `FileLinkFormatter` from the ErrorHandler component instead
+ * Add argument `$buildDir` to `WarmableInterface`
 
 6.3
 ---